evaluation/Hypothesis/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
apptools==5.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
bleach==4.1.0
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
configobj==5.0.6
contextvars==2.4
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
easydict==1.9
entrypoints==0.4
envisage==6.1.0
frozenlist==1.2.0
idna==3.3
idna-ssl==1.1.0
immutables==0.18
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.2
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.10
jsonschema==4.0.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-packaging==0.10.6
jupyter-server==1.13.1
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.1.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mayavi==4.8.0
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.2.0
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
ninja==1.10.2.3
notebook==6.4.10
numpy @ file:///C:/ci/numpy_and_numpy_base_1603480701039/work
open3d==0.15.1
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
pycparser==2.21
pyface==7.4.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
pytorch-metric-learning==1.5.2
pytz==2022.2.1
pywin32==304
pywinpty==1.1.6
pyzmq==23.2.1
requests==2.27.1
scikit-learn==0.24.2
scipy @ file:///C:/ci/scipy_1597675683670/work
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.2.0
tensorboardX==2.5.1
terminado==0.13.0
testpath==0.6.0
threadpoolctl==3.1.0
tomlkit==0.11.4
torch==1.10.2
torchvision @ file:///D:/Install_pack/torchvision-0.4.0%2Bcpu-cp36-cp36m-win_amd64.whl
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
traits==6.4.1
traitsui==7.4.0
typing_extensions==4.1.1
urllib3==1.26.12
vtk==9.1.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
widgetsnbextension==3.6.1
wincertstore==0.2
wslink==1.8.2
yarl==1.7.2
zipp==3.6.0
